refacto(store): implementing the container StateHandler
